Change Azure functions Implementation due to rest services being generated as Controllers
