+++ dotnet/src/extensions/Azure/Handlers/GeneXus.Deploy.AzureFunctions.Handlers.csproj
@@ -58,7 +58,7 @@
 	</ItemGroup>
 
 	<ItemGroup Condition="'$(HttpSupport)' == 'true'">
-		<PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http" Version="3.1.0" />
+		<PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http" Version="3.2.0" />
 		<Compile Include="HttpHandler\*" />
 	</ItemGroup>
 
@@ -69,6 +69,7 @@
 
 	<ItemGroup>
 		<PackageReference Include="Azure.Core" Version="1.42.0" />
+		<PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http.AspNetCore" Version="1.3.2" />
 		<PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk" Version="1.17.4" OutputItemType="Analyzer" />
 		<PackageReference Include="Microsoft.Azure.Functions.Worker" Version="1.23.0" />
 		<PackageReference Include="StackExchange.Redis" Version="2.6.122" />
@@ -99,6 +100,7 @@
 	</ItemGroup>
 	<ItemGroup>
 	  <Folder Include="Functions\" />
+	  <Folder Include="HttpHandler\" />
 	</ItemGroup>
 	<Target Name="CustomContentTarget">
 		<ItemGroup Condition="$(GenStdVersion)==''">
